I do not understand the activity to be carried out
I suppose that I have to create a test, so I consulted online and downloaded a format, I also saw a video of how to play poker and discovered that the code to write should not only distribute cards but also a map should be assigned. As they have not yet indicated the code to write, then I did the test on the algorithm. Attachment algorithm file